Organized Project to packages
