+++ run.py
@@ -1,0 +1,5 @@
+from Flask_blog import app
+
+
+if __name__ == '__main__':
+    app.run(port=5000, debug=True)
